refactor(models): merge thesis.py InvestmentThesis into unified_models.py
InvestmentThesis was the only ORM class in backend/models/thesis.py and it
already imported Base from unified_models. Folding it into the canonical
source eliminates a one-class module and standardizes import paths.

- Append InvestmentThesis to unified_models.py (preserves __tablename__,
  columns, indexes, relationships verbatim).
- Add `timezone` to the datetime import in unified_models.py (thesis used
  datetime.now(timezone.utc) defaults).
- Re-export InvestmentThesis from backend/models/__init__.py.
- Rewrite 5 callers:
    backend/api/routers/thesis.py
    backend/repositories/thesis_repository.py
    backend/tests/test_thesis_api.py
    backend/tests/integration/test_phase3_integration.py
    tests/security/test_row_locking.py
- Delete backend/models/thesis.py.

Smoke `python -c "from backend.models import InvestmentThesis; import backend.api.main"` returns OK.

Part of Tier B1 (docs/ANALYSIS_2026-05-14.md).
